Bitcoin Converter
This Java program was written for my Principles of Programming 2 class
during Fall 2013.  It was designed to find the current value of bitcoin
by parsing the HTML of bitcoinity.org and finding the current value of
(at the time) the three major exchanges.  The user would select the
desired exchange, input how many bitcoin they owned, and the program
would give them the current value of their bitcoin in USD.